added AnchorCollection & refactored change updates
