+++ Mapping Tools/Mapping_Tools.csproj
@@ -356,6 +356,7 @@
     <Compile Include="Components\Domain\ParsableDoubleListValidationRule.cs" />
     <Compile Include="Components\Domain\DoubleToStringConverter.cs" />
     <Compile Include="Components\DragAndDropListBox.cs" />
+    <Compile Include="Components\Graph\AnchorCollection.cs" />
     <Compile Include="Components\Graph\GraphIntegralDoubleAnimation.cs" />
     <Compile Include="Components\Graph\GraphDoubleAnimation.cs" />
     <Compile Include="Components\Graph\Interpolation\CustomIntegralExtremaAttribute.cs" />
